io: move posix process into a separate header
